requirements.txt
django==3.2.7
python-dotenv==0.19.0
sentry-sdk==1.4.3
django-requestlogging==1.0.1
djangorestframework==3.12.4